# feat(storage): separate key and values in imm to reduce allocation