--- a/src/storage/Cargo.toml
+++ b/src/storage/Cargo.toml
@@ -139,5 +139,9 @@ harness = false
 name = "bench_row"
 harness = false
 
+[[bench]]
+name = "bench_imm_compact"
+harness = false
+
 [lints]
 workspace = true