--- a/src/storage/src/hummock/shared_buffer/shared_buffer_batch.rs
+++ b/src/storage/src/hummock/shared_buffer/shared_buffer_batch.rs
@@ -288,7 +288,11 @@ impl SharedBufferBatch {
         *self.inner.epochs.last().unwrap()
     }
 
-    pub fn kv_count(&self) -> usize {
+    pub fn key_count(&self) -> usize {
+        self.inner.entries.len()
+    }
+
+    pub fn value_count(&self) -> usize {
         self.inner.new_values.len()
     }