@wenym1 2024-03-04
> the new structure LGTM, and I think you may conduct a perf test to check the improvement.

added a benchmark

main
```
bench-imm-merge/single-epoch time:   [261.48 ms 262.47 ms 263.49 ms]
bench-imm-merge/multi-epoch   time:   [337.32 ms 339.80 ms 343.92 ms]
```

this branch
```
bench-imm-merge/single-epoch time:   [204.69 ms 205.04 ms 205.48 ms]
bench-imm-merge/multi-epoch   time:   [326.13 ms 326.91 ms 327.72 ms]

```